% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_files.R
\name{add_files}
\alias{add_files}
\alias{add_files<-}
\alias{get_files}
\title{Add List of Files from Additional Table}
\usage{
add_files(refs, priority = "main text") <- value

get_files(refs)
}
\arguments{
\item{refs}{A data frame including reference entries imported from BibTeX
databases by \code{\link{read_bib}}. At least a column called 'bibtexkey'
have to be included in this table.}

\item{priority}{Character value. A keyword used as file description to define
which is the main document in the entry. Files including this description
will be listed first in the output bibtex file (JabRef style).}

\item{value}{A data frame listing files with respective bibtexkey and
MIME-Type. In this table the columns 'bibtexkey', 'file', and 'mime' are
mandatory. The occurrence of all bibtexkey values in 'value' will be
cross-checked and evenctually cause an error, thus you may clean this
data frame before using it.}
}
\value{
In \verb{add_files<-} the same data frame 'refs' with an inserted or updated
column 'file'.
For \code{get_files}, a data frame with three columns, namely 'bibtexkey', 'file',
'mime', and 'description' where every file is listed in aseparated row.
}
\description{
The function \verb{add_files<-} attempts to harmonize relational databases listing
files in a separated relation (data frame).

The reverse function is called \code{get_files}, where a string will be converted
into a data frame.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
