% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2lib_db.R
\name{db2lib_db}
\alias{db2lib_db}
\alias{db2lib_db.PostgreSQLConnection}
\title{Read bibliographic databases from PostgreSQL}
\usage{
db2lib_db(conn, ...)

\method{db2lib_db}{PostgreSQLConnection}(conn, schema, file_folder, simplify = FALSE, ...)
}
\arguments{
\item{conn}{A connection retrieved either by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
or \code{\link[divDB:connect_db]{divDB::connect_db()}}.}

\item{...}{Further arguments passed among methods.}

\item{schema}{A character value with the name of the schema containing the
main table and the file list. If the length is 2, the first value will be
assumed as the schema for the main table and the second, for the file
list. If the length is 1, the value will be recycled.}

\item{file_folder}{A character value showing the path to the linked documents
(i.e. PDF files of publications).}

\item{simplify}{Logical value indicating whether empty columns should be
skipped from output or not.}
}
\value{
A \linkS4class{lib_db} object.
}
\description{
Databases tabulated in PostgreSQL will be imported into a \linkS4class{lib_db}
object.
}
