% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list.R
\name{file_list}
\alias{file_list}
\alias{file_list.lib_df}
\alias{file_list<-}
\alias{file_list<-,lib_df,data.frame-method}
\title{Description of files associated to entries in BibTeX databases}
\usage{
file_list(x, ...)

\method{file_list}{lib_df}(x, ...)

file_list(x, ...) <- value

\S4method{file_list}{lib_df,data.frame}(x, priority = "main text", ...) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{lib_df}, usually reference entries
imported from a BibTeX databases by \code{\link[=read_bib]{read_bib()}}.}

\item{...}{Further parameters passed among methods.}

\item{value}{A data frame listing files with respective bibtexkey and
MIME-Type. In this table the columns 'bibtexkey', 'file', and 'mime' are
mandatory. The occurrence of all bibtexkey values in 'value' will be
cross-checked and eventually retrieve an error message.}

\item{priority}{Character value. A keyword used as file description to define
which is the main document in the entry. Files including this description
will be listed first in the output bibtex file (JabRef style).}
}
\value{
A data frame by \code{file_list()} or a \linkS4class{lib_df} object by \verb{file_list<-}.
}
\description{
Main documents as well as supplementary material may be linked to the
respective bibliographic entries in a BibTeX database. File names, MIME-types
and descriptions are stored in a single column by JabRef, separating columns
by colons and multiple files by semicolons. In a relational database, file
information may be stored in a separated table and linked by bibtexkeys.

The function \code{file_list()} produce this table, while the replace method
\verb{file_list<-} will use such table to write the column 'file' in a
\linkS4class{lib_df} object.
}
