% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_keys.R
\name{match_keys}
\alias{match_keys}
\alias{match_keys.character}
\alias{match_keys.lib_df}
\title{Match bibtexkeys used in an r-markdown document}
\usage{
match_keys(x, ...)

\method{match_keys}{character}(x, rmd_file, ...)

\method{match_keys}{lib_df}(x, rmd_file, ...)
}
\arguments{
\item{x}{An object of class 'lib_df'.}

\item{...}{Further arguments passed to \code{\link[=readLines]{readLines()}}.}

\item{rmd_file}{A character vector indicating the path to the r-markdown file
to be scanned.}
}
\value{
A data frame with two columns, \code{bibtexkey} for the found keys and \code{line}
with the line number of the occurrence of the key in the document.
}
\description{
This function scan an Rmd file in relation to a reference bibtex library
(the one used to insert citations) and finds the line position of the
respective citation.

This function can be useful to produce subsets of bibtex libraries to be
shared in companion to a depending r-markdown document.
It can be also used to count the occurrences of citations in a document and
through this way to detect rarely cited articles.
}
