% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaccess-data.R
\docType{data}
\name{openaccess-data}
\alias{openaccess-data}
\alias{openaccess}
\title{A BibTeX database with three open access publications}
\format{
An object of class \link[biblio:lib_df-class]{biblio::lib_df}.
}
\usage{
openaccess
}
\description{
Three open access publications extracted from
\href{https://bdj.pensoft.net/}{Biodiversity Data Journal} as example of a
BibTeX database with associated PDF files.
}
\examples{
data(openaccess)
openaccess
}
\keyword{datasets}
