% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.lib_df}
\title{Print content of lib_df objects}
\usage{
\method{print}{lib_df}(x, ...)
}
\arguments{
\item{x}{An object of class 'lib_df'.}

\item{...}{Further arguments passed among methods.}
}
\description{
A method for a brief overview on the content of a 'lib_df' object.
}
\examples{
Bib <- read_bib(bib=file.path(path.package("biblio"),
    "LuebertPliscoff.bib"))
Bib

}
\author{
Miguel Alvarez
}
