% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{print_comp}
\alias{print_comp}
\title{Print output of comparison}
\usage{
print_comp(comp)
}
\arguments{
\item{comp}{A list. The output generated by the function \code{compare_df()}.}
}
\description{
Report changes after update tables.
}
\keyword{internal}
