% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bib.R
\name{read_bib}
\alias{read_bib}
\title{Read BibTeX Databases}
\usage{
read_bib(bib, ...)
}
\arguments{
\item{bib}{Path to BibTeX file.}

\item{...}{Further arguments passed to \code{\link{readLines}}.

#' @examples
Bib <- read_bib(bib=file.path(path.package("biblio"),
"LuebertPliscoff.bib"))
Bib}
}
\description{
Reading BibTeX databases and importing into R as a data frame. All the fields
will be inserted as character values.
}
