% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pg.R
\name{read_pg}
\alias{read_pg}
\alias{read_pg.PostgreSQLConnection}
\title{Read bibliographic databases from PostgreSQL}
\usage{
read_pg(conn, ...)

\method{read_pg}{PostgreSQLConnection}(
  conn,
  name,
  main_table = "main_table",
  file_list = "file_list",
  add_files = TRUE,
  simplify = FALSE,
  ...
)
}
\arguments{
\item{conn}{A database connection established with \code{\link[=dbConnect]{dbConnect()}} or
\code{\link[dbaccess:connect_db]{dbaccess::connect_db2()}}.}

\item{...}{Further arguments passed among methods.}

\item{name}{Name of the database corresponding to a schema in the PostgreSQL
database.}

\item{main_table}{A character value indicating the name of the main table in
in the schema 'name'.}

\item{file_list}{A character value indicating the name of the table in schema
'name', which contains the names of the files and the respective
attributes.}

\item{add_files}{Logical value indicating whether information in table
'file_list' should be appended to the output or not.}

\item{simplify}{Logical value indicating whether empty columns should be
skipped from output or not.}
}
\value{
A \linkS4class{lib_df} object.
}
\description{
Databases tabulated in PostgreSQL will be imported into a \linkS4class{lib_df}
object.
}
