% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{sql_delete}
\alias{sql_delete}
\title{Write query and execute it for deleted entries}
\usage{
sql_delete(conn, obj, name, key, ...)
}
\arguments{
\item{conn}{A connection established by \code{\link[=dbConnect]{dbConnect()}}.}

\item{obj}{A comparison object created by \code{compare_df()}.}

\item{name}{A character vector with the names of schema and table in
PostgreSQL.}

\item{key}{A character value. The name of the primary key in the table.}

\item{...}{Further arguments (not yet used).}
}
\description{
Write query and execute it for deleted entries
}
\keyword{internal}
