% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\alias{update.PostgreSQLConnection}
\title{Compare Bibtex-files with Postgres databases and update}
\usage{
\method{update}{PostgreSQLConnection}(
  object,
  revision,
  key = "bibtexkey",
  name,
  db_args = list(),
  delete = FALSE,
  add = FALSE,
  update = FALSE,
  main_table = "main_table",
  file_list = "file_list",
  ...
)
}
\arguments{
\item{object}{A connection to a reference database established by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{revision}{A \linkS4class{lib_df} object imported by \code{\link[=read_bib]{read_bib()}}. This data
set represent an updated version of 'object'.}

\item{key}{A character value indicating the name of the column used as
identifier for references.}

\item{name}{Character value indicating the name of the schema in Postgres.
This argument is passed to \code{\link[=read_pg]{read_pg()}}.}

\item{db_args}{List of named arguments passed to \code{\link[=read_pg]{read_pg()}}.}

\item{delete}{Logical value indicating whether missing entries in 'bib' have
to be deleted in 'db'.}

\item{add}{Logical value indicating whether new entries in 'bib' have to be
inserted in 'db'.}

\item{update}{Logical value indicating whether entries modified in 'bib' have
to be updated in 'db'.}

\item{main_table}{Character value indicating the name of main table in
Postgres (see \code{\link[=read_pg]{read_pg()}}).}

\item{file_list}{Character value indicating the name of file list table in
Postgres (see \code{\link[=read_pg]{read_pg()}}).}

\item{...}{Further arguments passed to \code{\link[=pgInsert]{pgInsert()}}.}
}
\description{
When changes done in the Bibtex duplicated file, they can be briefly
displayed before export.

In \code{update_pg} actions 'delete', 'add', and 'update' have to be accordingly
set as \code{TRUE}, otherwise only \code{print_report()} will be executed.
}
