% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pg.R
\name{update_pg}
\alias{update_pg}
\alias{update_report}
\alias{update_report,lib_df,lib_df-method}
\alias{update_pg,PostgreSQLConnection,lib_df-method}
\alias{update_pg,PostgreSQLConnection,character-method}
\title{Compare Bibtex-files with Postgres databases and update}
\usage{
update_report(db, bib, ...)

\S4method{update_report}{lib_df,lib_df}(db, bib, get_files = TRUE, print_only = TRUE, ...)

update_pg(db, bib, ...)

\S4method{update_pg}{PostgreSQLConnection,lib_df}(
  db,
  bib,
  name,
  db_args = list(),
  delete = FALSE,
  add = FALSE,
  update = FALSE,
  main_table = "main_table",
  file_list = "file_list",
  ...
)

\S4method{update_pg}{PostgreSQLConnection,character}(db, bib, name, db_args = list(), bib_args = list(), ...)
}
\arguments{
\item{db}{Either a data frame imported by \code{\link[=read_pg]{read_pg()}} or a connection
established by \code{\link[=dbConnect]{dbConnect()}} to a reference database.}

\item{bib}{Either a data frame imported by \code{\link[=read_bib]{read_bib()}} or a path to a bibtex
file. This data set represent an updated version of 'db'.}

\item{...}{Further arguments passed among methods.}

\item{get_files}{Logical value indicating whether a list of files should be
extracted from 'bib' or not. If \code{TRUE}, then function \code{\link[=get_files]{get_files()}} will
be applied.}

\item{print_only}{Logical value indicating whether the outcome will be only
printed in the console or stored as a list.}

\item{name}{Character value indicating the name of the schema in Postgres.
This argument is passed to \code{\link[=read_pg]{read_pg()}}.}

\item{db_args}{List of named arguments passed to \code{\link[=read_pg]{read_pg()}}.}

\item{delete}{Logical value indicating whether missing entries in 'bib' have
to be deleted in 'db'.}

\item{add}{Logical value indicating whether new entries in 'bib' have to be
inserted in 'db'.}

\item{update}{Logical value indicating whether entries modified in 'bib' have
to be updated in 'db'.}

\item{main_table}{Character value indicating the name of main table in
Postgres (see \code{\link[=read_pg]{read_pg()}}).}

\item{file_list}{Character value indicating the name of file list table in
Postgres (see \code{\link[=read_pg]{read_pg()}}).}

\item{bib_args}{List of named arguments passed to \code{\link[=read_bib]{read_bib()}}.}
}
\description{
When changes done in the Bibtex duplicated file, they can be briefly
displayed before export.

In \code{update_pg} actions 'delete', 'add', and 'update' have to be accordingly
set as \code{TRUE}, otherwise only \code{print_report()} will be executed.
}
