% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{valid_file}
\alias{valid_file}
\title{Validation of files in database}
\usage{
valid_file(folder, df, ...)
}
\arguments{
\item{folder}{A character value indicating the path to the folder containing
reference files. Usually the folder set as \emph{Main file directory} in the
JabRef options.}

\item{df}{A data frame containing the file list (see \code{\link[=get_files]{get_files()}}).}

\item{...}{Further arguments (not yet used).}
}
\description{
List of local files are stored in an external folder. This function check the
existence of the mentioned files and also detect files stored in the folder
but not mentioned in the database.
}
\keyword{internal}
