% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bib.R
\name{write_bib}
\alias{write_bib}
\title{Write BibTeX Files}
\usage{
write_bib(obj, file, encoding = "UTF-8", ...)
}
\arguments{
\item{obj}{A data frame with bibliographic entries.}

\item{file}{A character value with the path and the name of the file to be
written.}

\item{encoding}{Character value with the encoding (passed to
\code{\link{file}}).}

\item{...}{Further arguments passed to \code{\link{file}}.}
}
\description{
BibTeX databases can be created from data frames, interacting with Postgres
databases.
}
