% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pg.R
\name{write_pg}
\alias{write_pg}
\title{Writing References from Data Frame to a Postgres Table}
\usage{
write_pg(
  conn,
  name,
  df1,
  df2,
  main_table = "main_table",
  file_list = "file_list",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{conn}{A connection established with \code{\link[=dbConnect]{dbConnect()}}.}

\item{name}{A character value with the name of the schema.}

\item{df1}{A data frame with the content for the main table.}

\item{df2}{A data frame with the content for the file table.}

\item{main_table}{A character value indicating the name of the main table in
the database.}

\item{file_list}{A character value indicating the name of the file list in
the database.}

\item{overwrite}{A logical value indicating whether existing tables should be
overwritten or not.}

\item{...}{Further arguments passed to pgInsert.}
}
\description{
This function implement the creation of a Postgres table and its population
with data imported by \code{\link[=read_bib]{read_bib()}}. To import data from a Postgres
table created by this function just use \code{\link[=dbReadTable]{dbReadTable()}}.
}
