% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2database.R
\name{bib2database}
\alias{bib2database}
\alias{bib2database,PostgreSQLConnection,character,missing-method}
\alias{bib2database,PostgreSQLConnection,character,lib_db-method}
\alias{bib2database,PostgreSQLConnection,character,lib_df-method}
\title{Create a database to store bibliographic references}
\usage{
bib2database(conn, schema, bib, ...)

\S4method{bib2database}{PostgreSQLConnection,character,missing}(conn, schema, comment = "", ...)

\S4method{bib2database}{PostgreSQLConnection,character,lib_db}(conn, schema, bib, ...)

\S4method{bib2database}{PostgreSQLConnection,character,lib_df}(conn, schema, bib, ...)
}
\arguments{
\item{conn}{A database connection.}

\item{schema}{A character value with the name of the schema where tables
will get stored. If the schema does not exist in the database, it will
be created by this function.}

\item{bib}{A \linkS4class{lib_df} or \linkS4class{lib_db} object. If missing, an empty
database will be created.}

\item{...}{Further arguments passed among methods.}

\item{comment}{A character value with the comment (description) assigned to
the created schema.}
}
\description{
A relational data model to create a storage of bibliographic references
(electronic library) from \linkS4class{lib_df} objects.
}
