% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{check_files}
\alias{check_files}
\alias{check_files.lib_db}
\title{Check for file existence}
\usage{
check_files(obj, ...)

\method{check_files}{lib_db}(obj, ...)
}
\arguments{
\item{obj}{A \linkS4class{lib_db} object including the path.}

\item{...}{Further arguments passed among methods.}
}
\description{
Documents and other associated files are assumed in JabRef to exist in a
single folder (here indicated by 'path') and this function cross-check their
occurrence.
}
