% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{check_files}
\alias{check_files}
\alias{check_files.lib_df}
\alias{check_files.PostgreSQLConnection}
\title{Check for file existence}
\usage{
check_files(x, ...)

\method{check_files}{lib_df}(x, path, ...)

\method{check_files}{PostgreSQLConnection}(x, path, name, db_args = list(), ...)
}
\arguments{
\item{x}{Either a \linkS4class{lib_df} object or a \linkS4class{PostgreSQLConnection}
including a variable listing associated files.}

\item{...}{Further arguments passed to \code{\link[=list.files]{list.files()}}.}

\item{path}{A character values with the path of the folder containing the
files, usually the documents as PDF files.}

\item{name}{Character value indicating the name of the schema in the
PostgreSQL database. It is passed to \code{\link[=read_pg]{read_pg()}}.}

\item{db_args}{A list with named arguments passed to \code{\link[=read_pg]{read_pg()}}.}
}
\description{
Documents and other associated files are assumed in JabRef to exist in a
single folder (here indicated by 'path') and this function cross-check their
occurrence.
}
