% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{compare_df}
\alias{compare_df}
\title{Compare data frames}
\usage{
compare_df(df1, df2, key)
}
\arguments{
\item{df1}{The reference data frame.}

\item{df2}{The updated data frame.}

\item{key}{A character value with the name of the variable used as primary
key in the tables.}
}
\description{
Detect changes between two versions of a data frame.
}
\keyword{internal}
